move get_dbt_multi_asset_args to the same file as dbt_assets
branch-name: move-get_dbt_multi_asset_args
